See previous commit for rational of deleting module
