+++ lib/simplecov/source_file.rb
@@ -6,7 +6,6 @@ module SimpleCov
   # source lines and featuring helpers to interpret that data.
   #
   class SourceFile
-    include SimpleCov::Supports::SourceFileSupport
     # Representation of a single line in a source file including
     # this specific line's source code, line_number and code coverage,
     # with the coverage being either nil (coverage not applicable, e.g. comment
@@ -78,8 +77,6 @@ module SimpleCov
     # Representing single branch that has been detected in coverage report.
     # Give us support methods that handle needed calculations.
     class Branch
-      include SimpleCov::Supports::BranchSupport
-
       attr_reader :type,
                   :id,
                   :start_line,
@@ -167,6 +164,73 @@ module SimpleCov
       #
       def badge
         positive? ? "+" : "-"
+      end
+
+      #
+      # Return array of sub branches of current branches
+      #
+      # @param [Array] branches
+      #
+      # @return [Array]
+      #
+      def sub_branches(branches)
+        if type == :case
+          case_branches(branches)
+        else
+          branches.select do |branch|
+            id == branch.root_id
+          end
+        end
+      end
+
+      #
+      # Case/When/Else have special branches behave because
+      # else branch is always present no matter if we declare it in code or not
+      # => Else declared start_line != root_branch_line
+      # => Else not declared start_line == root_branch_line
+      # for this case we ignore if its not declared becase it's useless
+      #
+      # @param [Array] branches
+      #
+      # @return [Array] <description>
+      #
+      def case_branches(branches)
+        branches.select do |branch|
+          id == branch.root_id && branch.start_line != start_line
+        end
+      end
+
+      # TODO: Refactoring candidate.
+      # Return boolean value depends on branches start_line & end_line
+      # including single branch conditions & nested branches, ex:
+      #   unless/if cond
+      #      puts "x"
+      #   end
+      #   if cond
+      #    ..
+      #   elsif cond
+      #    ..
+      #   else
+      #    ..
+      #   end
+      #
+      # @return [Boolean]
+      #
+      def inline_branch?(branches)
+        # nested conditions
+        return true if branches.any? { |b| b.root_id != id && sub_branches(branches).map(&:start_line).include?(b.start_line) }
+
+        # inline or single branch conditions
+        sub_branches(branches).all? { |branch| branch.start_line.eql?(start_line) && branch.end_line.eql?(end_line) }
+      end
+
+      #
+      # Return array with coverage count and badge
+      #
+      # @return [Array]
+      #
+      def report
+        [coverage, badge]
       end
     end
